backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
torch==2.1.2
numpy==1.26.3
python-multipart==0.0.6
python-dotenv==1.0.0
pytorch-crf==0.7.2
khmercut==0.1.0